Add the ability to temporarily change video resolution in options menu.
Occasionally I encountered streams and videos using none standard resolutions, or resolutions that are too high with the "Chunked" setting for playback on my device (PS3). So instead of having backing all the way out to get the the plugin settings I added this.

This is not a perfect fix. I would like to make it possible to read and write the settings permanently from the options menu to the main plugin settings, but I have yet to find a way to do this. To be clear, the setting this commit adds only lasts for as long as you are on the page you set it for, the resolution option in the main plugin settings dialog is essentially unused following this commit. The default setting is 720p30 which is supported by Twitch most of the time and plays back smoothly on PS3. If anyone knows a better way to achieve this, pull requests are welcomed.