Show native resolution of live streams
This patch also changes the default video quality setting back to chunked, which play properly on ps3 for most streams. The only thing that doesn't seem to play properly is 1080p60 or higher. With this patch you'll be able to see the resolution before you connect and change it if you so desire. When I see Native Resolution: 1080p60, I can set 720p60 and not have frames dropped or audio desync.